+++ include/jsoncons_ext/jsonpointer/jsonpointer.hpp
@@ -1268,7 +1268,7 @@ namespace jsonpointer {
 
         switch (parent_value.type())
         {
-            case json_type::array_value:
+            case json_type::array:
             {
                 if (parent_value.empty())
                 {
@@ -1290,7 +1290,7 @@ namespace jsonpointer {
                 break;
             }
 
-            case json_type::object_value:
+            case json_type::object:
             {
                 if (parent_value.empty())
                 {